refactor (handle XML): move XML processing logic to utils
